[ALPS02359542] tb6735p1_64: build warning
[Detail]
build warning "implicitly truncated" happen when pre-release

[Solution]
add (u32) before sizeof function in gt9xx_update.c

[Feature] Others

MTK-Commit-Id: 2effe52595cc4c782069cdbe9c957fe815390968

CR-Id: ALPS02359542
(cherry picked from commit ab86ca64a40eb3ff2ace29ca75353feab4fa6164)

+++ drivers/input/touchscreen/mediatek/GT9XXTB_hotknot/gt9xx_update.c
@@ -652,7 +652,7 @@ static u8 gup_check_update_file(struct i2c_client *client, struct st_fw_head *fw
 #if defined(CONFIG_GTP_AUTO_UPDATE)
 #if defined(CONFIG_GTP_HEADER_FW_UPDATE)
 		GTP_INFO("Update by default firmware array");
-		update_msg.fw_total_len = sizeof(gtp_default_FW) - FW_HEAD_LENGTH;
+		update_msg.fw_total_len = (u32)(sizeof(gtp_default_FW) - FW_HEAD_LENGTH);
 		if (sizeof(gtp_default_FW) <
 		    (FW_HEAD_LENGTH + FW_SECTION_LENGTH * 4 + FW_DSP_ISP_LENGTH + FW_DSP_LENGTH +
 		     FW_BOOT_LENGTH)) {
